requirements.txt
biopython==1.75
numpy==1.17.4
pandas==0.25.3
scipy==1.4.1
python-igraph==0.8.2